requirements.txt
flask==2.3.2
flask-socketio==5.3.3
nextcord==2.4.2
peewee==3.16.2
progress==1.6
python-dotenv==1.0.0
